localized majority of playwright tests
