+++ apps/webapp/tests-playwright/local/sae-bench.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-test('should navigate to the SAE Evals page', async ({ page }) => {
+test.skip('should navigate to the SAE Evals page', async ({ page }) => {
   await page.goto('/');
   await page.click('text=SAE Evals');
   await expect(page).toHaveURL('/sae-bench');
